+++ src/pytorch_lightning/loops/epoch/training_epoch_loop.py
@@ -108,15 +108,13 @@ class TrainingEpochLoop(loops.Loop[_OUTPUTS_TYPE]):
         if self.trainer.should_stop:
             # early stopping
             min_epochs = self.trainer.fit_loop.min_epochs
-            met_min_epochs = self.trainer.current_epoch >= min_epochs if min_epochs else True
-            met_min_steps = self.global_step >= self.min_steps if self.min_steps else True
-            met_min_conditions = met_min_epochs and met_min_steps
-            if not met_min_conditions:
+            should_stop_early = self.trainer.fit_loop._should_stop_early
+            if not should_stop_early:
                 self._warning_cache.info(
                     f"Trainer was signaled to stop but the required `min_epochs={min_epochs!r}` or"
                     f" `min_steps={self.min_steps!r}` has not been met. Training will continue..."
                 )
-            return met_min_conditions
+            return should_stop_early
 
         return False